Добавила авторизационные методы
