+++ src/services/actions/registration.ts
@@ -1,6 +1,6 @@
 import { RegisterUser } from '../../shared/models/register-user.type';
 import { ActionType } from '../../shared/models/action.type';
-import { registerUser } from '../../shared/api/user.service';
+import { registerUser } from '../../shared/api/auth.service';
 
 export const REGISTRATION = 'REGISTRATION';
 export const REGISTRATION_REQUEST = 'REGISTRATION_REQUEST';
