+++ src/shared/models/login-user.type.ts
@@ -1,0 +1,4 @@
+export type LoginUser = {
+  password: string;
+  email: string;
+};
